fix: FFI「外」模块与终端 VT 处理做平台隔离,macOS/Linux 可构建
- evaluator: 拆出 ffi_windows.go(Windows 原 DLL 逻辑)与 ffi_other.go(其他平台返回暂不支持),封装 ffiLoadDLL/ffiCall,evaluator.go 三处改用封装函数
- main: 移除 enableVirtualTerminalProcessing 的 syscall 实现,拆为 term_windows.go / term_other.go(非 Windows 空实现)
- 修复 macOS/Linux 上 go build 报 undefined syscall (LoadDLL/DLL/GetConsoleMode 等)

Closes #3
